Validator+LLVM: Treat match/is result as actual type rather than pointer
Previously match/is would hand you a pointer (with appropriate
mutability) to the variant. The logic behind this was that this is
consistent with what is actually happening behind the scenes.

However this is less useful and rather inconvenient so let's just
treat the result as the actual variant type. I justify it by pointing
at how pointer dereference has the same semantics wrt "giving" you
the actual type while still being a implicit pointer semantically.