Validator: Fix function specialization order dependence
Gets rid of a nasty hack in a nicer way. I checked this change on main
to ensure validity as this branch does not current pass the tests due to
throwing away messages. As a bonus is technically faster by a couple
tens of ms as if that matters right now lol

Most of this patch's bulk is just changes to randomize file order and
handle that without making dump index order assumptions