Validator: Prevent rare race condition
A race condition can occur if one thread attempts to create a new
specialization while another thread is filling the shape and
pre-existing specializations where the newly created specialization
will not be filled.

We avoid that by preventing any threads from creating specializations
while another thread is actively filling the shape and pre-existing
specializations. However it must still allow that same thread to access
the shape deeper in the callstack as it is valid in some cases for a
user type to be self-referential, eg via a pointer, so it must be able
to add new specializations mid-filling.